+++ src/utils.cpp
@@ -2,8 +2,8 @@
 
 #include "utils.h"
 
-double exponential_random(double lambda) {
-    double u;
+float exponential_random(float lambda) {
+    float u;
     u = rand() / (RAND_MAX + 1.0);
 
     return -log(1 - u) / lambda;
