+++ src/utils.h
@@ -1,6 +1,6 @@
 #ifndef COMPUTER_NETWORKS_UTILS_H
 #define COMPUTER_NETWORKS_UTILS_H
 
-double exponential_random(double lambda);
+float exponential_random(float lambda);
 
 #endif //COMPUTER_NETWORKS_UTILS_H
